requirements.txt
# Core Dependencies
mcp>=1.0.0
facebook-business>=18.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Async Support
asyncio>=3.4.3

# Utilities
pydantic>=2.0.0
typing-extensions>=4.0.0
aiohttp>=3.9.0  # For async HTTP requests
